Fixing method initialization, separate tests for CubeWeightsBuilder and make constructor protected
